# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "FAQ Management System",
        "title": "FAQ API",
        "contact": {
            "name": "Ahmad Abdelrazik",
            "email": "ahmad.abdelrazik.swe@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/faq-categories": {
            "get": {
                "description": "Retrieve a list of all existing FAQ categories",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FAQ Categories"
                ],
                "summary": "Get all FAQ categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.getAllFaqCategoryResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new FAQ category (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FAQ Categories"
                ],
                "summary": "Create FAQ category",
                "parameters": [
                    {
                        "description": "Category Name",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.createFaqCategoryInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.createFaqCategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Unauthorized/Admin only",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Category already exists",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/faq-categories/{category}": {
            "get": {
                "description": "Retrieve all global FAQs for a specific category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Global FAQs"
                ],
                "summary": "Get global FAQs by category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category Name",
                        "name": "category",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.getAllGlobalFaqsResponse"
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the name of an existing FAQ category (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FAQ Categories"
                ],
                "summary": "Update FAQ category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Current Category Name",
                        "name": "category",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New Category Name",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.updateFaqCategoryInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.updateFaqCategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Name conflict or edit conflict",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new global FAQ within a category (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Global FAQs"
                ],
                "summary": "Create global FAQ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category Name",
                        "name": "category",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "FAQ Data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.createFaqInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.createFaqResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove an FAQ category from the system (Admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FAQ Categories"
                ],
                "summary": "Delete FAQ category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category Name",
                        "name": "category",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.deletedFaqCategoryResponse"
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.createFaqCategoryInput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "controllers.createFaqCategoryResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "controllers.createFaqInput": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                }
            }
        },
        "controllers.createFaqResponse": {
            "type": "object",
            "properties": {
                "faq": {
                    "$ref": "#/definitions/controllers.faqDTO"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "controllers.deletedFaqCategoryResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "controllers.faqCategoryDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "controllers.faqDTO": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                }
            }
        },
        "controllers.getAllFaqCategoryResponse": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.faqCategoryDTO"
                    }
                }
            }
        },
        "controllers.getAllGlobalFaqsResponse": {
            "type": "object",
            "properties": {
                "faqs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.faqDTO"
                    }
                }
            }
        },
...
        "controllers.updateFaqCategoryInput": {
            "type": "object",
            "properties": {
                "new_name": {
                    "type": "string"
                }
            }
        },
        "controllers.updateFaqCategoryResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        }
    }
}